src/requirements.txt
gunicorn
pandas 
numpy
flask
flask_restful
flask-cors 
h5py
pytorch